+++ src/apis/index.ts
@@ -6,5 +6,5 @@ export const BACKEND_URL = process.env.NEXT_PUBLIC_BASE_URL as string;
 export const instance = axios.create({
   baseURL: BACKEND_URL,
   withCredentials: true,
-  // headers: { Authorization: `Bearer ${여기에 토큰이 들어갑니다.}` },
+  headers: { Authorization: `Bearer $"{여기에 토큰이 들어갑니다."}` },
 });
